feat(api): seeding for env
