+++ appeals/api/src/database/seed/LPAs/prod.js
@@ -1,0 +1,12 @@
+/** @typedef {import('#utils/db-client/index.js').Prisma.LPACreateInput} LPA */
+
+/**
+ * @returns {LPA[]}
+ */
+export const localPlanningDepartmentList = [
+	{
+		lpaCode: 'Q9999',
+		name: 'System Test Borough Council',
+		email: 'appealplanningdecisiontest@planninginspectorate.gov.uk'
+	}
+];
