+++ appeals/api/src/database/seed/data-samples.js
@@ -103,54 +103,6 @@ export const agentsList = personList.map(({ firstName, lastName, company, email 
 	email
 }));
 
-export const localPlanningDepartmentList = [
-	{
-		lpaCode: 'Q9999',
-		name: 'Integration Testing Borough Council',
-		email: 'int2@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'MAID',
-		name: 'Maidstone Borough Council',
-		email: 'maid@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'BARN',
-		name: 'Barnsley Metropolitan Borough Council',
-		email: 'barn@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'WORT',
-		name: 'Worthing Borough Council',
-		email: 'wort@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'DORS',
-		name: 'Dorset Council',
-		email: 'dors@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'BASI',
-		name: 'Basingstoke and Deane Borough Council',
-		email: 'basi@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'WILT',
-		name: 'Wiltshire Council',
-		email: 'wilt@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'WAVE',
-		name: 'Waveney District Council',
-		email: 'wave@lpa-email.gov.uk'
-	},
-	{
-		lpaCode: 'BRIS',
-		name: 'Bristol City Council',
-		email: 'bris@lpa-email.gov.uk'
-	}
-];
-
 /** @type {AppealSite[]} */
 export const addressesList = [
 	{
