--- azure-pipelines-db-seed-appeals-production.yml
@@ -1,55 +1,0 @@
-parameters:
-  - name: environment
-    displayName: Environment
-    type: string
-    default: Prod
-    values:
-      - Prod
-
-pool: pins-odt-agent-pool
-
-pr: none
-
-trigger: none
-
-resources:
-  repositories:
-    - repository: templates
-      type: github
-      endpoint: Planning-Inspectorate
-      name: Planning-Inspectorate/common-pipeline-templates
-      ref: refs/tags/release/3.16.1
-
-variables:
-  - template: variables/environments/${{ lower(parameters.environment )}}.yml@templates
-  - group: pipeline_secrets
-
-jobs:
-  - deployment: Seed_DB
-    displayName: Seed ${{ parameters.environment }} Database
-    environment: ${{ parameters.environment }}
-    strategy:
-      runOnce:
-        deploy:
-          steps:
-            - download: none
-            - checkout: self
-            - template: steps/node_script.yml@templates
-              parameters:
-                script: npm ci
-                nodeVersion: 18
-            - template: steps/azure_auth.yml@templates
-            - template: steps/azure_get_secrets.yml@templates
-              parameters:
-                keyVaultName: pins-kv-appeals-bo-prod
-                secrets:
-                  - name: appeals-bo-sql-admin-connection-string
-                    variable: appeals-bo-sql-admin-connection-string
-            - template: steps/node_script.yml@templates
-              parameters:
-                environmentVariables:
-                  DATABASE_URL: $(appeals-bo-sql-admin-connection-string)
-                script: npm run db:seed
-                workingDirectory: $(Build.Repository.LocalPath)/appeals/api
-    workspace:
-      clean: all
