+++ azure-pipelines-db-seed-appeals.yml
@@ -6,6 +6,8 @@ parameters:
     values:
       - Dev
       - Test
+      - Training
+      - Prod
 
 pool: pins-odt-agent-pool
 
@@ -40,17 +42,44 @@ jobs:
                 script: npm ci
                 nodeVersion: 18
             - template: steps/azure_auth.yml@templates
+            # kv name is different in the training env (had to be shorter)
+            - bash: |
+                if [[ $ENVIRONMENT == "training" ]]; then
+                  echo "Name: pins-kv-appeals-bo-train (env: $ENVIRONMENT)"
+                  echo "##vso[task.setvariable variable=kvName;]pins-kv-appeals-bo-train"
+                else
+                  echo "Name: pins-kv-appeals-bo-$ENVIRONMENT (env: $ENVIRONMENT)"
+                  echo "##vso[task.setvariable variable=kvName;]pins-kv-appeals-bo-$ENVIRONMENT"
+                fi
+              displayName: Set KV name
+              env:
+                ENVIRONMENT: $(ENVIRONMENT)
             - template: steps/azure_get_secrets.yml@templates
               parameters:
-                keyVaultName: pins-kv-appeals-bo-${{ lower(parameters.environment) }}
+                keyVaultName: $(kvName)
                 secrets:
                   - name: appeals-bo-sql-admin-connection-string
                     variable: appeals-bo-sql-admin-connection-string
             - template: steps/node_script.yml@templates
               parameters:
+                condition: and(ne('${{ parameters.environment }}', 'Prod'), ne('${{ parameters.environment }}', 'Training'))
                 environmentVariables:
                   DATABASE_URL: $(appeals-bo-sql-admin-connection-string)
                 script: npm run db:seed
+                workingDirectory: $(Build.Repository.LocalPath)/appeals/api
+            - template: steps/node_script.yml@templates
+              parameters:
+                condition: and(eq('${{ parameters.environment }}', 'Training'))
+                environmentVariables:
+                  DATABASE_URL: $(appeals-bo-sql-admin-connection-string)
+                script: npm run db:seed:training
+                workingDirectory: $(Build.Repository.LocalPath)/appeals/api
+            - template: steps/node_script.yml@templates
+              parameters:
+                condition: and(eq('${{ parameters.environment }}', 'Prod'))
+                environmentVariables:
+                  DATABASE_URL: $(appeals-bo-sql-admin-connection-string)
+                script: npm run db:seed:prod
                 workingDirectory: $(Build.Repository.LocalPath)/appeals/api
     workspace:
       clean: all
